Rename files to wp recommendations
